CoreRT-specific Object.cs changes - Fix build breaks - Apply same cleanup to all related files
