+++ src/Runtime.Base/src/System/Object.cs
@@ -2,26 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-/*============================================================
-**
-**
-**
-** Object is the root class for all CLR objects.  This class
-** defines only the basics.
-**
-** 
-===========================================================*/
-
-using System.Runtime;
-using System.Diagnostics;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
+using Internal.Runtime;
 using Internal.Runtime.CompilerServices;
-
-// TODO: remove when m_pEEType becomes EETypePtr
-using EEType = Internal.Runtime.EEType;
-using ObjHeader = Internal.Runtime.ObjHeader;
 
 namespace System
 {
@@ -30,20 +14,11 @@ namespace System
     // Data Contract: Single field of type EEType*
     // VTable Contract: The first vtable slot should be the finalizer for object => The first virtual method in the object class should be the Finalizer
 
-    // The Object is the root class for all object in the CLR System. Object 
-    // is the super class for all other CLR objects and provide a set of methods and low level
-    // services to subclasses. 
-
     public unsafe class Object
     {
         // CS0649: Field '{blah}' is never assigned to, and will always have its default value
 #pragma warning disable 649
-
-        // Marked as internal for now so that some classes can use C#'s fixed statement on objects. 
-        // Wouldn't have to do this if we could directly declared pinned locals.
-        // TODO: Consider making this EETypePtr instead of EEType*.
-        internal EEType* m_pEEType;
-
+        private EEType* m_pEEType;
 #pragma warning restore
 
         // Creates a new instance of an Object.
@@ -60,7 +35,7 @@ namespace System
         {
         }
 
-        internal unsafe EEType* EEType
+        internal EEType* EEType
         {
             get
             {
@@ -68,7 +43,6 @@ namespace System
                 //        unsafe.  But, generically, we don't expect managed code such as this to be allowed
                 //        to run while the GC is running.
                 return m_pEEType;
-                //PREFER m_pEEType.ToPointer();
             }
         }
